Add atomic CAS support to CompareAndSwapOp
This adds an overlay for JDK 11 to enable CAS instruction support
in the CompareAndSwapOp. It uses the UseLSE flag to determine whether
to use the CAS instruction or not.
